+++ js/form/upload-photo.js
